cmd/simon: Fix issue with server watcher loop cancelation
The old version had to either observe a tick or job request before
looking at the cancelation signal. If someone set a high tick
interval this would could block the server exit for a while.

Now, the watcher looks at the cancelation signal twice. First at
the beginning of an iteration (this ensures cancelation has higher
priority). Second, concurently with waiting for a tick or job
request. This ensures a long tick interval won't block the loop.

The worker loop is unaffected as it was already doing the right
thing.